backend-cc/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.1
tzlocal==5.3.1
whitenoise==6.9.0
dotenv
requests